requirements.txt
appdirs==1.4.4
attrs==20.3.0
bcrypt==3.2.0
black==20.8b1
cffi==1.14.4
click==7.1.2
cryptography==3.3.1
forwardable==0.4.1
future==0.18.2
importlib-metadata==3.3.0
importlib-resources==3.3.0
iniconfig==1.1.1
mccabe==0.6.1
mock==4.0.3
mypy-extensions==0.4.3
netmiko==3.4.0
ntc-templates==1.6.0
packaging==20.8
paramiko==2.7.2
pathspec==0.8.1
pluggy==0.13.1
py==1.10.0
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==6.1.1
pyflakes==2.3.1
pylama==7.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyserial==3.5
pytest==6.2.0
pytest-mock==3.5.1
PyYAML==5.3.1
regex==2020.11.13
scp==0.13.3
six==1.15.0
snowballstemmer==2.1.0
tenacity==6.2.0
textfsm==1.1.0
toml==0.10.2
typed-ast==1.4.2
typing-extensions==3.7.4.3
zipp==3.4.0
